+++ docs/src/changelog.md
@@ -1,5 +1,6 @@
 # ChangeLog of NNHelferlein package
 
++ print_network changed to summary
 + Pretrained nets saved at zenodo and simplified constructors added
 + AbstractNN and AbstractLayer added
 + copy model and save/load as JLD2 added